refactor(user creation): move password logic to the user entity
