+++ server/entity/User.ts
@@ -10,6 +10,11 @@ import {
 import { Permission, hasPermission } from '../lib/permissions';
 import { MediaRequest } from './MediaRequest';
 import bcrypt from 'bcrypt';
+import path from 'path';
+import PreparedEmail from '../lib/email';
+import logger from '../logger';
+import { getSettings } from '../lib/settings';
+import { default as generatePassword } from 'secure-random-password';
 
 @Entity()
 export class User {
@@ -85,5 +90,38 @@ export class User {
         return reject(false);
       }
     });
+  }
+
+  public async setPassword(password: string): Promise<void> {
+    const hashedPassword = await bcrypt.hash(password, 12);
+    this.password = hashedPassword;
+  }
+
+  public async resetPassword(): Promise<void> {
+    const password = generatePassword.randomPassword({ length: 16 });
+    this.setPassword(password);
+
+    const applicationUrl = getSettings().main.applicationUrl;
+    try {
+      logger.info(`Sending password email for ${this.email}`, {
+        label: 'User creation',
+      });
+      const email = new PreparedEmail();
+      await email.send({
+        template: path.join(__dirname, '../templates/email/password'),
+        message: {
+          to: this.email,
+        },
+        locals: {
+          password: password,
+          applicationUrl,
+        },
+      });
+    } catch (e) {
+      logger.error('Failed to send out password email', {
+        label: 'User creation',
+        message: e.message,
+      });
+    }
   }
 }
