+++ server/routes/user.ts
@@ -6,11 +6,7 @@ import { User } from '../entity/User';
 import { hasPermission, Permission } from '../lib/permissions';
 import { getSettings } from '../lib/settings';
 import logger from '../logger';
-import bcrypt from 'bcrypt';
 import gravatarUrl from 'gravatar-url';
-import { default as generatePassword } from 'secure-random-password';
-import path from 'path';
-import PreparedEmail from '../lib/email';
 
 const router = Router();
 
@@ -27,53 +23,26 @@ router.post('/', async (req, res, next) => {
     const body = req.body;
     const userRepository = getRepository(User);
 
-    // if not passing an explicit password, generate a secure one
     const passedExplicitPassword = body.password && body.password.length > 0;
-    const password = passedExplicitPassword
-      ? body.password
-      : generatePassword.randomPassword({ length: 16 });
-
-    const hashedPassword = await bcrypt.hash(password, 12);
     const avatar = gravatarUrl(body.email);
 
     const user = new User({
       avatar: body.avatar ?? avatar,
       username: body.username ?? body.email,
       email: body.email,
-      password: hashedPassword,
+      password: body.password,
       permissions: body.permissions,
       plexToken: '',
       userType: body.userType,
     });
 
-    await userRepository.save(user);
-
-    if (!passedExplicitPassword) {
-      const applicationUrl = getSettings().main.applicationUrl;
-      try {
-        logger.info(`Sending password email for ${body.email}`, {
-          label: 'User creation',
-        });
-        // const email = getNewEmail();
-        const email = new PreparedEmail();
-        await email.send({
-          template: path.join(__dirname, '../templates/email/password'),
-          message: {
-            to: body.email,
-          },
-          locals: {
-            password: password,
-            applicationUrl,
-          },
-        });
-      } catch (e) {
-        logger.error('Failed to send out password email', {
-          label: 'User creation',
-          message: e.message,
-        });
-      }
+    if (passedExplicitPassword) {
+      await user.setPassword(body.password);
+    } else {
+      await user.resetPassword();
     }
 
+    await userRepository.save(user);
     return res.status(201).json(user.filter());
   } catch (e) {
     next({ status: 500, message: e.message });
